+++ services/index.tsx
@@ -90,3 +90,18 @@ export const getSimilarPosts = async () => {
 
   return result.posts
 }
+
+export const getCategories = async () => {
+  const query = gql`
+    query GetCategories {
+      categories {
+        name
+        slug
+      }
+    }
+  `
+  if (!graphqlAPI) return
+  const result = await request(graphqlAPI, query)
+
+  return result.categories
+}
